pyproject.toml
[project]
name = "agent-safe"
version = "0.1.0"
description = "A governance and policy enforcement layer for AI agents and non-human identities"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "Agent-Safe Contributors" }]
keywords = [
    "ai", "agents", "governance", "policy", "security",
    "kubernetes", "audit", "compliance", "nhi", "non-human-identity",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "pyjwt>=2.8,<3.0",
    "click>=8.0,<9.0",
]

[project.urls]
Homepage = "https://github.com/sahb4k/agent-safe"
Documentation = "https://github.com/sahb4k/agent-safe/tree/main/docs"
Repository = "https://github.com/sahb4k/agent-safe"
Issues = "https://github.com/sahb4k/agent-safe/issues"
Changelog = "https://github.com/sahb4k/agent-safe/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
s3 = [
    "boto3>=1.28",
]

[project.scripts]
agent-safe = "agent_safe.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_safe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = ["ignore::jwt.InsecureKeyLengthWarning"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["SIM102"]  # nested if-not-None guard pattern is clearer than single-line
